Disable SyncFileRange on arm
syscall.SyncFileRange is not defined on arm (though it is for arm64?).

Fix a warning about a comparison that is always true when ints are
32-bits.

Fixes #171

+++ internal/arenaskl/node.go
@@ -65,11 +65,11 @@ func newNode(
 		panic("height cannot be less than one or greater than the max height")
 	}
 	keySize := key.Size()
-	if keySize > math.MaxUint32 {
+	if int64(keySize) > math.MaxUint32 {
 		panic("key is too large")
 	}
 	valueSize := len(value)
-	if len(value) > math.MaxUint32 {
+	if int64(len(value)) > math.MaxUint32 {
 		panic("value is too large")
 	}
